Add actor management: implement actor creation, retrieval, and position updating in the database; update game state handling for actor integration

+++ server/pkg/packets/constructors.go
@@ -46,8 +46,8 @@ func NewRegisterResponse(success bool, err error) Msg {
 func NewActorInfo(actor *objs.Actor) Msg {
 	return &Packet_ActorInfo{
 		ActorInfo: &ActorInfo{
-			X:    actor.X,
-			Y:    actor.Y,
+			X:    int32(actor.X),
+			Y:    int32(actor.Y),
 			Name: actor.Name,
 		},
 	}
